Refactor WooPosVariationsDataSource to utilize WooPosVariationsTypesFilterConfig for variation filtering
