+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/woopos/common/data/WooPosVariationsTypesFilterConfig.kt
